@@ -6,19 +6,19 @@ homepage: https://github.com/daohoangson/flutter_widget_from_html/tree/master/pa
 environment:
   # Stack.clipBehavior https://github.com/flutter/flutter/commit/42e02d60d3104e97161ba597103854557b2d72ee
   flutter: ">=1.20.0 <2.0.0"
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  html: ^0.14.0+3
+  html: ^0.15.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  golden_toolkit: ^0.6.0
-  network_image_mock: ^1.0.1
-  pedantic: ^1.9.0
+  golden_toolkit: ^0.9.0-0
+  mocktail_image_network: ^0.0.1-0
+  pedantic: ^1.10.0
   test: any
 
 flutter:
