@@ -14,10 +14,11 @@ abstract class BuildMetadata {
   /// The associatd element.
   final dom.Element element;
 
-  final TextStyleBuilder _tsb;
+  /// The associated [TextStyleBuilder].
+  final TextStyleBuilder tsb;
 
   /// Creates a node.
-  BuildMetadata(this.element, this._tsb);
+  BuildMetadata(this.element, this.tsb);
 
   /// The registered build ops.
   Iterable<BuildOp> get buildOps;
@@ -42,14 +43,14 @@ abstract class BuildMetadata {
   /// - Inline style `display: block`
   ///
   /// See [BuildOp.onWidgetsIsOptional].
-  bool get willBuildSubtree;
+  bool? get willBuildSubtree;
 
   /// Adds an inline style.
   operator []=(String key, String value);
 
   /// Gets an inline style value by key.
-  String operator [](String key) {
-    String value;
+  String? operator [](String key) {
+    String? value;
     for (final x in styles) {
       if (x.key == key) value = x.value;
     }
@@ -60,17 +61,7 @@ abstract class BuildMetadata {
   void register(BuildOp op);
 
   @override
-  String toString() =>
-      'BuildMetadata(${element == null ? "root" : element.outerHtml})';
-
-  /// Enqueues a text style builder callback.
-  ///
-  /// Returns the associated [TextStyleBuilder].
-  TextStyleBuilder tsb<T>([
-    TextStyleHtml Function(TextStyleHtml tsh, T input) builder,
-    T input,
-  ]) =>
-      _tsb..enqueue(builder, input);
+  String toString() => 'BuildMetadata(${element.outerHtml})';
 }
 
 /// A building operation to customize how a DOM element is rendered.
@@ -99,7 +90,7 @@ class BuildOp {
   ///
   /// Note: op must be registered early for this to work e.g.
   /// in [WidgetFactory.parse] or [onChild].
-  final Map<String, String> Function(dom.Element element) defaultStyles;
+  final Map<String, String> Function(dom.Element element)? defaultStyles;
 
   /// The callback that will be called whenver a child element is found.
   ///
@@ -115,16 +106,16 @@ class BuildOp {
   /// );
   ///
   /// ```
-  final void Function(BuildMetadata childMeta) onChild;
+  final void Function(BuildMetadata childMeta)? onChild;
 
   /// The callback that will be called when child elements have been processed.
-  final void Function(BuildMetadata meta, BuildTree tree) onTree;
+  final void Function(BuildMetadata meta, BuildTree tree)? onTree;
 
   /// The callback that will be called when child elements have been built.
   ///
   /// Note: only works if it's a block element.
-  final Iterable<Widget> Function(
-      BuildMetadata meta, Iterable<WidgetPlaceholder> widgets) onWidgets;
+  final Iterable<Widget>? Function(
+      BuildMetadata meta, Iterable<WidgetPlaceholder> widgets)? onWidgets;
 
   /// Controls whether the element should be forced to be rendered as block.
   ///
