@@ -12,40 +12,39 @@ import 'core_html_widget.dart';
 
 /// A factory to build widgets.
 class WidgetFactory {
-  BuildOp _styleBgColor;
-  BuildOp _styleBlock;
-  BuildOp _styleBorder;
-  BuildOp _styleDisplayNone;
-  BuildOp _styleMargin;
-  BuildOp _stylePadding;
-  BuildOp _styleSizing;
-  BuildOp _styleTextDecoration;
-  BuildOp _styleVerticalAlign;
-  BuildOp _tagA;
-  BuildOp _tagBr;
-  BuildOp _tagFont;
-  BuildOp _tagHr;
-  BuildOp _tagImg;
-  BuildOp _tagPre;
-  BuildOp _tagQ;
-  TextStyleHtml Function(TextStyleHtml, String) __tsbFontSize;
-  TextStyleHtml Function(TextStyleHtml, String) _tsbLineHeight;
-  HtmlWidget _widget;
+  BuildOp? _styleBgColor;
+  BuildOp? _styleBlock;
+  BuildOp? _styleBorder;
+  BuildOp? _styleDisplayNone;
+  BuildOp? _styleMargin;
+  BuildOp? _stylePadding;
+  BuildOp? _styleSizing;
+  BuildOp? _styleTextDecoration;
+  BuildOp? _styleVerticalAlign;
+  BuildOp? _tagA;
+  BuildOp? _tagBr;
+  BuildOp? _tagFont;
+  BuildOp? _tagHr;
+  BuildOp? _tagImg;
+  BuildOp? _tagPre;
+  BuildOp? _tagQ;
+  TextStyleHtml Function(TextStyleHtml, String)? __tsbFontSize;
+  TextStyleHtml Function(TextStyleHtml, String)? _tsbLineHeight;
+  HtmlWidget? _widget;
 
   /// Builds [Align].
-  Widget buildAlign(
+  Widget? buildAlign(
           BuildMetadata meta, Widget child, AlignmentGeometry alignment) =>
-      alignment == null ? child : Align(alignment: alignment, child: child);
+      Align(alignment: alignment, child: child);
 
   /// Builds [AspectRatio].
-  Widget buildAspectRatio(
+  Widget? buildAspectRatio(
           BuildMetadata meta, Widget child, double aspectRatio) =>
-      aspectRatio == null
-          ? child
-          : AspectRatio(aspectRatio: aspectRatio, child: child);
+      AspectRatio(aspectRatio: aspectRatio, child: child);
 
   /// Builds primary column (body).
-  WidgetPlaceholder buildBody(BuildMetadata meta, Iterable<Widget> children) =>
+  WidgetPlaceholder? buildBody(
+          BuildMetadata meta, Iterable<WidgetPlaceholder> children) =>
       buildColumnPlaceholder(meta, children, trimMarginVertical: true);
 
   /// Builds [border] with [Container] or [DecoratedBox].
@@ -53,27 +52,25 @@ class WidgetFactory {
   /// See https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing
   /// for more information regarding `content-box` (the default)
   /// and `border-box` (set [isBorderBox] to use).
-  Widget buildBorder(BuildMetadata meta, Widget child, BoxBorder border,
+  Widget? buildBorder(BuildMetadata meta, Widget child, BoxBorder border,
           {bool isBorderBox = false}) =>
-      border == null
-          ? child
-          : isBorderBox == true
-              ? DecoratedBox(
-                  child: child,
-                  decoration: BoxDecoration(border: border),
-                )
-              : Container(
-                  child: child,
-                  decoration: BoxDecoration(border: border),
-                );
+      isBorderBox
+          ? DecoratedBox(
+              child: child,
+              decoration: BoxDecoration(border: border),
+            )
+          : Container(
+              child: child,
+              decoration: BoxDecoration(border: border),
+            );
 
   /// Builds column placeholder.
-  WidgetPlaceholder buildColumnPlaceholder(
+  WidgetPlaceholder? buildColumnPlaceholder(
     BuildMetadata meta,
-    Iterable<Widget> children, {
+    Iterable<WidgetPlaceholder> children, {
     bool trimMarginVertical = false,
   }) {
-    if (children?.isNotEmpty != true) return null;
+    if (children.isEmpty) return null;
 
     if (children.length == 1) {
       final child = children.first;
@@ -95,9 +92,9 @@ class WidgetFactory {
   }
 
   /// Builds [Column].
-  Widget buildColumnWidget(
+  Widget? buildColumnWidget(
       BuildMetadata meta, TextStyleHtml tsh, List<Widget> children) {
-    if (children?.isNotEmpty != true) return null;
+    if (children.isEmpty) return null;
     if (children.length == 1) return children.first;
 
     return Column(
@@ -109,53 +106,49 @@ class WidgetFactory {
   }
 
   /// Builds [DecoratedBox].
-  Widget buildDecoratedBox(
+  Widget? buildDecoratedBox(
     BuildMetadata meta,
     Widget child, {
-    Color color,
+    Color? color,
   }) =>
-      child != null
-          ? DecoratedBox(
-              child: child,
-              decoration: BoxDecoration(
-                color: color,
-              ),
-            )
-          : child;
+      DecoratedBox(
+        child: child,
+        decoration: BoxDecoration(
+          color: color,
+        ),
+      );
 
   /// Builds 1-pixel-height divider.
-  Widget buildDivider(BuildMetadata meta) => const DecoratedBox(
+  Widget? buildDivider(BuildMetadata meta) => const DecoratedBox(
         decoration: BoxDecoration(color: Color.fromRGBO(0, 0, 0, 1)),
         child: SizedBox(height: 1),
       );
 
   /// Builds [GestureDetector].
-  Widget buildGestureDetector(
+  Widget? buildGestureDetector(
           BuildMetadata meta, Widget child, GestureTapCallback onTap) =>
-      onTap == null ? child : GestureDetector(child: child, onTap: onTap);
+      GestureDetector(child: child, onTap: onTap);
 
   /// Builds horizontal scroll view.
-  Widget buildHorizontalScrollView(BuildMetadata meta, Widget child) =>
+  Widget? buildHorizontalScrollView(BuildMetadata meta, Widget child) =>
       SingleChildScrollView(child: child, scrollDirection: Axis.horizontal);
 
-  /// Builds [Image] from [provider].
-  Widget buildImage(BuildMetadata meta, Object provider, ImageMetadata data) {
-    if (provider == null) return null;
+  /// Builds image widget from an [ImageMetadata].
+  Widget? buildImage(BuildMetadata meta, ImageMetadata data) {
+    final src = data.sources.isNotEmpty ? data.sources.first : null;
+    if (src == null) return null;
 
-    Widget built;
-    if (provider is ImageProvider) {
-      final semanticLabel = data?.alt ?? data?.title;
-      built = Image(
-        errorBuilder: (_, error, __) {
-          print('$provider error: $error');
-          final text = semanticLabel ?? '❌';
-          return Text(text);
-        },
-        excludeFromSemantics: semanticLabel == null,
-        fit: BoxFit.fill,
-        image: provider,
-        semanticLabel: semanticLabel,
-      );
+    var built = buildImageWidget(
+      meta,
+      semanticLabel: data.alt ?? data.title,
+      url: src.url,
+    );
+
+    if (built != null &&
+        src.height?.isNegative == false &&
+        src.width?.isNegative == false &&
+        src.height != 0) {
+      built = buildAspectRatio(meta, built, src.width! / src.height!);
     }
 
     if (_widget?.onTapImage != null && built != null) {
@@ -166,14 +159,46 @@ class WidgetFactory {
     return built;
   }
 
+  /// Builds [Image].
+  Widget? buildImageWidget(
+    BuildMetadata meta, {
+    String? semanticLabel,
+    required String url,
+  }) {
+    late final ImageProvider? provider;
+    if (url.startsWith('asset:')) {
+      provider = imageProviderFromAsset(url);
+    } else if (url.startsWith('data:image/')) {
+      provider = imageProviderFromDataUri(url);
+    } else if (url.startsWith('file:')) {
+      provider = imageProviderFromFileUri(url);
+    } else {
+      provider = imageProviderFromNetwork(url);
+    }
+    if (provider == null) return null;
+
+    return Image(
+      errorBuilder: (_, error, __) {
+        print('$provider error: $error');
+        final text = semanticLabel ?? '❌';
+        return Text(text);
+      },
+      excludeFromSemantics: semanticLabel == null,
+      fit: BoxFit.fill,
+      image: provider,
+      semanticLabel: semanticLabel,
+    );
+  }
+
   /// Builds [Padding].
-  Widget buildPadding(BuildMetadata meta, Widget child, EdgeInsets padding) =>
-      padding == null || padding == const EdgeInsets.all(0)
+  Widget? buildPadding(
+          BuildMetadata meta, Widget child, EdgeInsetsGeometry padding) =>
+      padding == EdgeInsets.zero
           ? child
           : Padding(child: child, padding: padding);
 
   /// Builds [Stack].
-  Widget buildStack(
+  Widget? buildStack(
           BuildMetadata meta, TextStyleHtml tsh, List<Widget> children) =>
       Stack(
         children: children,
@@ -182,24 +207,23 @@ class WidgetFactory {
       );
 
   /// Builds [RichText].
-  Widget buildText(BuildMetadata meta, TextStyleHtml tsh, InlineSpan text) =>
+  Widget? buildText(BuildMetadata meta, TextStyleHtml tsh, InlineSpan text) =>
       RichText(
-        overflow: tsh?.textOverflow ?? TextOverflow.clip,
+        overflow: tsh.textOverflow ?? TextOverflow.clip,
         text: text,
-        textAlign: tsh?.textAlign ?? TextAlign.start,
-        textDirection: tsh?.textDirection ?? TextDirection.ltr,
+        textAlign: tsh.textAlign ?? TextAlign.start,
+        textDirection: tsh.textDirection,
 
         // TODO: calculate max lines automatically for ellipsis if needed
         // currently it only renders 1 line with ellipsis
-        maxLines: tsh?.maxLines == -1 ? null : tsh?.maxLines,
+        maxLines: tsh.maxLines == -1 ? null : tsh.maxLines,
       );
 
   /// Prepares [GestureTapCallback].
-  GestureTapCallback gestureTapCallback(String url) => url != null
-      ? () => _widget.onTapUrl != null
-          ? _widget.onTapUrl(url)
-          : print('[HtmlWidget] onTapUrl($url)')
-      : null;
+  GestureTapCallback? gestureTapCallback(String url) =>
+      () => _widget?.onTapUrl != null
+          ? _widget!.onTapUrl!(url)
+          : print('[HtmlWidget] onTapUrl($url)');
 
   /// Returns [context]-based dependencies.
   ///
@@ -272,7 +296,7 @@ class WidgetFactory {
     return '';
   }
 
-  String _getListStyleMarkerRoman(int i) {
+  String? _getListStyleMarkerRoman(int i) {
     // TODO: find some lib to generate programatically
     const map = <int, String>{
       1: 'I',
@@ -290,55 +314,38 @@ class WidgetFactory {
     return map[i];
   }
 
-  /// Returns [ImageProvider].
-  Object imageProvider(ImageSource imgSrc) {
-    if (imgSrc == null) return null;
-    final url = imgSrc.url;
-
-    if (url.startsWith('asset:')) {
-      return _imageFromAsset(url);
-    }
-
-    if (url.startsWith('data:')) {
-      return _imageFromDataUri(url);
-    }
-
-    if (url.startsWith('file:')) {
-      return _imageFromFileUri(url);
-    }
-
-    return _imageFromUrl(url);
-  }
-
-  Object _imageFromAsset(String url) {
-    final uri = url?.isNotEmpty == true ? Uri.tryParse(url) : null;
+  /// Returns an [AssetImage].
+  ImageProvider? imageProviderFromAsset(String url) {
+    final uri = Uri.parse(url);
     final assetName = uri.path;
-    if (assetName?.isNotEmpty != true) return null;
+    if (assetName.isEmpty) return null;
 
-    final package = uri.queryParameters?.containsKey('package') == true
+    final package = uri.queryParameters.containsKey('package') == true
         ? uri.queryParameters['package']
         : null;
 
     return AssetImage(assetName, package: package);
   }
 
-  Object _imageFromDataUri(String dataUri) {
+  /// Returns a [MemoryImage].
+  ImageProvider? imageProviderFromDataUri(String dataUri) {
     final bytes = bytesFromDataUri(dataUri);
     if (bytes == null) return null;
 
     return MemoryImage(bytes);
   }
 
-  Object _imageFromFileUri(String url) {
-    final uri = url?.isNotEmpty == true ? Uri.tryParse(url) : null;
-    final filePath = uri?.toFilePath();
-    if (filePath?.isNotEmpty != true) return null;
+  /// Returns a [FileImage].
+  ImageProvider? imageProviderFromFileUri(String url) {
+    final filePath = Uri.parse(url).toFilePath();
+    if (filePath.isEmpty) return null;
 
     return FileImage(File(filePath));
   }
 
-  Object _imageFromUrl(String url) =>
-      url?.isNotEmpty == true ? NetworkImage(url) : null;
+  /// Returns a [NetworkImage].
+  ImageProvider? imageProviderFromNetwork(String url) =>
+      url.isNotEmpty ? NetworkImage(url) : null;
 
   /// Prepares the root [TextStyleBuilder].
   void onRoot(TextStyleBuilder rootTsb) {}
@@ -350,12 +357,12 @@ class WidgetFactory {
     switch (meta.element.localName) {
       case kTagA:
         _tagA ??= TagA(this, () => _widget?.hyperlinkColor).buildOp;
-        meta.register(_tagA);
+        meta.register(_tagA!);
         break;
 
       case 'abbr':
       case 'acronym':
-        meta.tsb(
+        meta.tsb.enqueue(
           TextStyleOps.textDeco,
           TextDeco(style: TextDecorationStyle.dotted, under: true),
         );
@@ -364,7 +371,7 @@ class WidgetFactory {
       case 'address':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
-          ..tsb(TextStyleOps.fontStyle, FontStyle.italic);
+          ..tsb.enqueue(TextStyleOps.fontStyle, FontStyle.italic);
         break;
 
       case 'article':
@@ -388,16 +395,16 @@ class WidgetFactory {
 
       case 'b':
       case 'strong':
-        meta.tsb(TextStyleOps.fontWeight, FontWeight.bold);
+        meta.tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
 
       case 'big':
-        meta.tsb(_tsbFontSize, kCssFontSizeLarger);
+        meta.tsb.enqueue(_tsbFontSize, kCssFontSizeLarger);
         break;
 
       case 'br':
         _tagBr ??= BuildOp(onTree: (_, tree) => tree.addNewLine());
-        meta.register(_tagBr);
+        meta.register(_tagBr!);
         break;
 
       case kTagCenter:
@@ -411,14 +418,15 @@ class WidgetFactory {
       case 'em':
       case 'i':
       case 'var':
-        meta.tsb(TextStyleOps.fontStyle, FontStyle.italic);
+        meta.tsb.enqueue(TextStyleOps.fontStyle, FontStyle.italic);
         break;
 
       case kTagCode:
       case kTagKbd:
       case kTagSamp:
       case kTagTt:
-        meta.tsb(TextStyleOps.fontFamily, [kTagCodeFont1, kTagCodeFont2]);
+        meta.tsb
+            .enqueue(TextStyleOps.fontFamily, [kTagCodeFont1, kTagCodeFont2]);
         break;
       case kTagPre:
         _tagPre ??= BuildOp(
@@ -432,7 +440,7 @@ class WidgetFactory {
         );
         meta
           ..[kCssDisplay] = kCssDisplayBlock
-          ..register(_tagPre);
+          ..register(_tagPre!);
         break;
 
       case 'dd':
@@ -446,89 +454,92 @@ class WidgetFactory {
       case 'dt':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
 
       case 'del':
       case 's':
       case 'strike':
-        meta.tsb(TextStyleOps.textDeco, TextDeco(strike: true));
+        meta.tsb.enqueue(TextStyleOps.textDeco, TextDeco(strike: true));
         break;
 
       case kTagFont:
         _tagFont ??= BuildOp(
           defaultStyles: (element) {
             final attrs = element.attributes;
+            final color = attrs[kAttributeFontColor];
+            final fontFace = attrs[kAttributeFontFace];
+            final fontSize = kCssFontSizes[attrs[kAttributeFontSize] ?? ''];
             return {
-              kCssColor: attrs[kAttributeFontColor],
-              kCssFontFamily: attrs[kAttributeFontFace],
-              kCssFontSize: kCssFontSizes[attrs[kAttributeFontSize]],
+              if (color != null) kCssColor: color,
+              if (fontFace != null) kCssFontFamily: fontFace,
+              if (fontSize != null) kCssFontSize: fontSize,
             };
           },
         );
-        meta.register(_tagFont);
+        meta.register(_tagFont!);
         break;
 
       case 'hr':
         _tagHr ??= BuildOp(
           defaultStyles: (_) => const {'margin-bottom': '1em'},
-          onWidgets: (meta, _) => [buildDivider(meta)],
+          onWidgets: (meta, _) => listOrNull(buildDivider(meta)),
         );
         meta
           ..[kCssDisplay] = kCssDisplayBlock
-          ..register(_tagHr);
+          ..register(_tagHr!);
         break;
 
       case 'h1':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
           ..[kCssMargin] = '0.67em 0'
-          ..tsb(_tsbFontSize, '2em')
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(_tsbFontSize, '2em')
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
       case 'h2':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
           ..[kCssMargin] = '0.83em 0'
-          ..tsb(_tsbFontSize, '1.5em')
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(_tsbFontSize, '1.5em')
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
       case 'h3':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
           ..[kCssMargin] = '1em 0'
-          ..tsb(_tsbFontSize, '1.17em')
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(_tsbFontSize, '1.17em')
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
       case 'h4':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
           ..[kCssMargin] = '1.33em 0'
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
       case 'h5':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
           ..[kCssMargin] = '1.67em 0'
-          ..tsb(_tsbFontSize, '0.83em')
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(_tsbFontSize, '0.83em')
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
       case 'h6':
         meta
           ..[kCssDisplay] = kCssDisplayBlock
           ..[kCssMargin] = '2.33em 0'
-          ..tsb(_tsbFontSize, '0.67em')
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(_tsbFontSize, '0.67em')
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
 
       case kTagImg:
         _tagImg ??= TagImg(this).buildOp;
-        meta.register(_tagImg);
+        meta.register(_tagImg!);
         break;
 
       case 'ins':
       case 'u':
-        meta.tsb(TextStyleOps.textDeco, TextDeco(under: true));
+        meta.tsb.enqueue(TextStyleOps.textDeco, TextDeco(under: true));
         break;
 
       case kTagOrderedList:
@@ -552,7 +563,7 @@ class WidgetFactory {
 
       case kTagQ:
         _tagQ ??= TagQ(this).buildOp;
-        meta.register(_tagQ);
+        meta.register(_tagQ!);
         break;
 
       case kTagRuby:
@@ -565,18 +576,18 @@ class WidgetFactory {
         break;
 
       case 'small':
-        meta.tsb(_tsbFontSize, kCssFontSizeSmaller);
+        meta.tsb.enqueue(_tsbFontSize, kCssFontSizeSmaller);
         break;
 
       case 'sub':
         meta
           ..[kCssVerticalAlign] = kCssVerticalAlignSub
-          ..tsb(_tsbFontSize, kCssFontSizeSmaller);
+          ..tsb.enqueue(_tsbFontSize, kCssFontSizeSmaller);
         break;
       case 'sup':
         meta
           ..[kCssVerticalAlign] = kCssVerticalAlignSuper
-          ..tsb(_tsbFontSize, kCssFontSizeSmaller);
+          ..tsb.enqueue(_tsbFontSize, kCssFontSizeSmaller);
         break;
 
       case kTagTable:
@@ -595,7 +606,7 @@ class WidgetFactory {
       case kTagTableHeaderCell:
         meta
           ..[kCssVerticalAlign] = kCssVerticalAlignMiddle
-          ..tsb(TextStyleOps.fontWeight, FontWeight.bold);
+          ..tsb.enqueue(TextStyleOps.fontWeight, FontWeight.bold);
         break;
       case kTagTableCaption:
         meta[kCssTextAlign] = kCssTextAlignCenter;
@@ -620,35 +631,39 @@ class WidgetFactory {
       case kCssBackground:
       case kCssBackgroundColor:
         _styleBgColor ??= StyleBgColor(this).buildOp;
-        meta.register(_styleBgColor);
+        meta.register(_styleBgColor!);
         break;
 
       case kCssColor:
         final color = tryParseColor(value);
-        if (color != null) meta.tsb(TextStyleOps.color, color);
+        if (color != null) meta.tsb.enqueue(TextStyleOps.color, color);
         break;
 
       case kCssDirection:
-        meta.tsb(TextStyleOps.textDirection, value);
+        meta.tsb.enqueue(TextStyleOps.textDirection, value);
         break;
 
       case kCssFontFamily:
         final list = TextStyleOps.fontFamilyTryParse(value);
-        if (list != null) meta.tsb(TextStyleOps.fontFamily, list);
+        meta.tsb.enqueue(TextStyleOps.fontFamily, list);
         break;
 
       case kCssFontSize:
-        meta.tsb(_tsbFontSize, value);
+        meta.tsb.enqueue(_tsbFontSize, value);
         break;
 
       case kCssFontStyle:
         final fontStyle = TextStyleOps.fontStyleTryParse(value);
-        if (fontStyle != null) meta.tsb(TextStyleOps.fontStyle, fontStyle);
+        if (fontStyle != null) {
+          meta.tsb.enqueue(TextStyleOps.fontStyle, fontStyle);
+        }
         break;
 
       case kCssFontWeight:
         final fontWeight = TextStyleOps.fontWeightTryParse(value);
-        if (fontWeight != null) meta.tsb(TextStyleOps.fontWeight, fontWeight);
+        if (fontWeight != null) {
+          meta.tsb.enqueue(TextStyleOps.fontWeight, fontWeight);
+        }
         break;
 
       case kCssHeight:
@@ -658,18 +673,18 @@ class WidgetFactory {
       case kCssMinWidth:
       case kCssWidth:
         _styleSizing ??= StyleSizing(this).buildOp;
-        meta.register(_styleSizing);
+        meta.register(_styleSizing!);
         break;
 
       case kCssLineHeight:
         _tsbLineHeight ??= TextStyleOps.lineHeight(this);
-        meta.tsb(_tsbLineHeight, value);
+        meta.tsb.enqueue(_tsbLineHeight!, value);
         break;
 
       case kCssMaxLines:
       case kCssMaxLinesWebkitLineClamp:
         final maxLines = value == kCssMaxLinesNone ? -1 : int.tryParse(value);
-        if (maxLines != null) meta.tsb(TextStyleOps.maxLines, maxLines);
+        if (maxLines != null) meta.tsb.enqueue(TextStyleOps.maxLines, maxLines);
         break;
 
       case kCssTextAlign:
@@ -681,52 +696,54 @@ class WidgetFactory {
           for (final style in meta.styles) {
             if (style.key == kCssTextDecoration) {
               final textDeco = TextDeco.tryParse(style.values);
-              if (textDeco != null) meta.tsb(TextStyleOps.textDeco, textDeco);
+              if (textDeco != null) {
+                meta.tsb.enqueue(TextStyleOps.textDeco, textDeco);
+              }
             }
           }
         });
-        meta.register(_styleTextDecoration);
+        meta.register(_styleTextDecoration!);
         break;
 
       case kCssTextOverflow:
         switch (value) {
           case kCssTextOverflowClip:
-            meta.tsb(TextStyleOps.textOverflow, TextOverflow.clip);
+            meta.tsb.enqueue(TextStyleOps.textOverflow, TextOverflow.clip);
             break;
           case kCssTextOverflowEllipsis:
-            meta.tsb(TextStyleOps.textOverflow, TextOverflow.ellipsis);
+            meta.tsb.enqueue(TextStyleOps.textOverflow, TextOverflow.ellipsis);
             break;
         }
         break;
 
       case kCssVerticalAlign:
         _styleVerticalAlign ??= StyleVerticalAlign(this).buildOp;
-        meta.register(_styleVerticalAlign);
+        meta.register(_styleVerticalAlign!);
         break;
     }
 
     if (key.startsWith(kCssBorder)) {
       _styleBorder ??= StyleBorder(this).buildOp;
-      meta.register(_styleBorder);
+      meta.register(_styleBorder!);
     }
 
     if (key.startsWith(kCssMargin)) {
       _styleMargin ??= StyleMargin(this).buildOp;
-      meta.register(_styleMargin);
+      meta.register(_styleMargin!);
     }
 
     if (key.startsWith(kCssPadding)) {
       _stylePadding ??= StylePadding(this).buildOp;
-      meta.register(_stylePadding);
+      meta.register(_stylePadding!);
     }
   }
 
   /// Parses display inline style.
-  void parseStyleDisplay(BuildMetadata meta, String value) {
+  void parseStyleDisplay(BuildMetadata meta, String? value) {
     switch (value) {
       case kCssDisplayBlock:
         _styleBlock ??= DisplayBlockOp(this);
-        meta.register(_styleBlock);
+        meta.register(_styleBlock!);
         break;
       case kCssDisplayNone:
         _styleDisplayNone ??= BuildOp(
@@ -737,7 +754,7 @@ class WidgetFactory {
           },
           priority: 0,
         );
-        meta.register(_styleDisplayNone);
+        meta.register(_styleDisplayNone!);
         break;
       case kCssDisplayTable:
         meta.register(TagTable(this, meta).op);
@@ -748,13 +765,13 @@ class WidgetFactory {
   /// Resets for a new build.
   @mustCallSuper
   void reset(State state) {
-    final widget = state?.widget;
+    final widget = state.widget;
     _widget = widget is HtmlWidget ? widget : null;
   }
 
   /// Resolves full URL with [HtmlWidget.baseUrl] if available.
-  String urlFull(String url) {
-    if (url?.isNotEmpty != true) return null;
+  String? urlFull(String url) {
+    if (url.isEmpty) return null;
     if (url.startsWith('data:')) return url;
 
     final uri = Uri.tryParse(url);
@@ -768,7 +785,6 @@ class WidgetFactory {
   }
 
   TextStyleHtml Function(TextStyleHtml, String) get _tsbFontSize {
-    __tsbFontSize ??= TextStyleOps.fontSize(this);
-    return __tsbFontSize;
+    return __tsbFontSize ??= TextStyleOps.fontSize(this);
   }
 }
