@@ -28,12 +28,12 @@ class StyleSizing {
 
   BuildOp get buildOp => BuildOp(
         onTree: (meta, tree) {
-          if (meta.willBuildSubtree) return;
+          if (meta.willBuildSubtree == true) return;
 
           final input = _parse(meta);
           if (input == null) return;
 
-          WidgetPlaceholder widget;
+          WidgetPlaceholder? widget;
           for (final b in tree.bits) {
             if (b is WidgetBit) {
               if (widget != null) return;
@@ -43,25 +43,25 @@ class StyleSizing {
             }
           }
 
-          widget?.wrapWith((c, w) => _build(c, w, input, meta.tsb()));
+          widget?.wrapWith((c, w) => _build(c, w, input, meta.tsb));
         },
         onWidgets: (meta, widgets) {
-          if (!meta.willBuildSubtree) return widgets;
+          if (meta.willBuildSubtree == false) return widgets;
 
           final input = _parse(meta);
           if (input == null) return widgets;
           return listOrNull(wf
               .buildColumnPlaceholder(meta, widgets)
-              ?.wrapWith((c, w) => _build(c, w, input, meta.tsb())));
+              ?.wrapWith((c, w) => _build(c, w, input, meta.tsb)));
         },
         onWidgetsIsOptional: true,
         priority: kPriority5k,
       );
 
-  _StyleSizingInput _parse(BuildMetadata meta) {
-    CssLength maxHeight, maxWidth, minHeight, minWidth;
-    Axis preferredAxis;
-    CssLength preferredHeight, preferredWidth;
+  _StyleSizingInput? _parse(BuildMetadata meta) {
+    CssLength? maxHeight, maxWidth, minHeight, minWidth;
+    Axis? preferredAxis;
+    CssLength? preferredHeight, preferredWidth;
 
     for (final style in meta.styles) {
       switch (style.key) {
@@ -144,7 +144,7 @@ class StyleSizing {
     );
   }
 
-  static CssSizingValue _getValue(CssLength length, TextStyleHtml tsh) {
+  static CssSizingValue? _getValue(CssLength? length, TextStyleHtml tsh) {
     if (length == null) return null;
 
     final value = length.getValue(tsh);
@@ -163,13 +163,13 @@ class StyleSizing {
 
 @immutable
 class _StyleSizingInput {
-  final CssLength maxHeight;
-  final CssLength maxWidth;
-  final CssLength minHeight;
-  final CssLength minWidth;
-  final Axis preferredAxis;
-  final CssLength preferredHeight;
-  final CssLength preferredWidth;
+  final CssLength? maxHeight;
+  final CssLength? maxWidth;
+  final CssLength? minHeight;
+  final CssLength? minWidth;
+  final Axis? preferredAxis;
+  final CssLength? preferredHeight;
+  final CssLength? preferredWidth;
 
   _StyleSizingInput({
     this.maxHeight,
