@@ -30,33 +30,31 @@ class TagLi {
   final _itemMetas = <BuildMetadata>[];
   final _itemWidgets = <WidgetPlaceholder>[];
 
-  _ListConfig _config;
-  BuildOp _itemOp;
-  BuildOp _listOp;
+  _ListConfig? _config;
+  BuildOp? _itemOp;
+  BuildOp? _listOp;
 
   TagLi(this.wf, this.listMeta);
 
   _ListConfig get config {
     // cannot build config from constructor because
     // inline styles are not fully parsed at that time
-    _config ??= _ListConfig.fromBuildMetadata(listMeta);
-    return _config;
+    return _config ??= _ListConfig.fromBuildMetadata(listMeta);
   }
 
   BuildOp get op {
-    _listOp ??= _TagLiListOp(this);
-    return _listOp;
+    return _listOp ??= _TagLiListOp(this);
   }
 
   Map<String, String> defaultStyles(dom.Element element) {
     final attrs = element.attributes;
-    final depth = listMeta.parentOps?.whereType<_TagLiListOp>()?.length ?? 0;
+    final depth = listMeta.parentOps.whereType<_TagLiListOp>().length;
 
     final styles = {
       'padding-inline-start': '40px',
       kCssListStyleType: element.localName == kTagOrderedList
           ? _ListConfig.listStyleTypeFromAttributeType(
-                  attrs[kAttributeLiType]) ??
+                  attrs[kAttributeLiType] ?? '') ??
               kCssListStyleTypeDecimal
           : depth == 0
               ? kCssListStyleTypeDisc
@@ -87,17 +85,17 @@ class TagLi {
       },
     );
 
-    childMeta.register(_itemOp);
+    childMeta.register(_itemOp!);
   }
 
   Widget _buildItem(BuildContext context, Widget child, int i) {
     final meta = _itemMetas[i];
     final listStyleType = _ListConfig.listStyleTypeFromBuildMetadata(meta) ??
         config.listStyleType;
-    final markerIndex = config.markerReversed == true
+    final markerIndex = config.markerReversed
         ? (config.markerStart ?? _itemWidgets.length) - i
         : (config.markerStart ?? 1) + i;
-    final tsh = meta.tsb().build(context);
+    final tsh = meta.tsb.build(context);
 
     final markerText = wf.getListStyleMarker(listStyleType, markerIndex);
     final marker = _buildMarker(tsh, listStyleType, markerText);
@@ -111,7 +109,7 @@ class TagLi {
 
   Widget _buildMarker(TextStyleHtml tsh, String type, String text) {
     final style = tsh.styleWithHeight;
-    return text?.isNotEmpty == true
+    return text.isNotEmpty
         ? RichText(
             maxLines: 1,
             overflow: TextOverflow.clip,
@@ -131,11 +129,11 @@ class TagLi {
 class _ListConfig {
   final String listStyleType;
   final bool markerReversed;
-  final int markerStart;
+  final int? markerStart;
 
   _ListConfig({
-    this.listStyleType,
-    this.markerReversed,
+    required this.listStyleType,
+    required this.markerReversed,
     this.markerStart,
   });
 
@@ -149,15 +147,15 @@ class _ListConfig {
     );
   }
 
-  static String listStyleTypeFromBuildMetadata(BuildMetadata meta) {
+  static String? listStyleTypeFromBuildMetadata(BuildMetadata meta) {
     final listStyleType = meta[kCssListStyleType];
     if (listStyleType != null) return listStyleType;
 
     return listStyleTypeFromAttributeType(
-        meta.element.attributes[kAttributeLiType]);
+        meta.element.attributes[kAttributeLiType] ?? '');
   }
 
-  static String listStyleTypeFromAttributeType(String type) {
+  static String? listStyleTypeFromAttributeType(String type) {
     switch (type) {
       case kAttributeLiTypeAlphaLower:
         return kCssListStyleTypeAlphaLower;
@@ -181,10 +179,10 @@ class _ListItem extends MultiChildRenderObjectWidget {
   final TextDirection textDirection;
 
   _ListItem({
-    Widget child,
-    Key key,
-    Widget marker,
-    this.textDirection,
+    required Widget child,
+    Key? key,
+    required Widget marker,
+    required this.textDirection,
   }) : super(children: [child, marker], key: key);
 
   @override
@@ -211,7 +209,7 @@ class _ListItemRenderObject extends RenderBox
         ContainerRenderObjectMixin<RenderBox, _ListItemData>,
         RenderBoxContainerDefaultsMixin<RenderBox, _ListItemData> {
   _ListItemRenderObject({
-    TextDirection textDirection,
+    required TextDirection textDirection,
   }) : _textDirection = textDirection;
 
   TextDirection get textDirection => _textDirection;
@@ -223,27 +221,27 @@ class _ListItemRenderObject extends RenderBox
   }
 
   @override
-  double computeDistanceToActualBaseline(TextBaseline baseline) =>
+  double? computeDistanceToActualBaseline(TextBaseline baseline) =>
       defaultComputeDistanceToFirstActualBaseline(baseline);
 
   @override
   double computeMaxIntrinsicHeight(double width) =>
-      firstChild.computeMaxIntrinsicHeight(width);
+      firstChild!.computeMaxIntrinsicHeight(width);
 
   @override
   double computeMaxIntrinsicWidth(double height) =>
-      firstChild.computeMaxIntrinsicWidth(height);
+      firstChild!.computeMaxIntrinsicWidth(height);
 
   @override
   double computeMinIntrinsicHeight(double width) =>
-      firstChild.computeMinIntrinsicHeight(width);
+      firstChild!.computeMinIntrinsicHeight(width);
 
   @override
   double computeMinIntrinsicWidth(double height) =>
-      firstChild.getMinIntrinsicWidth(height);
+      firstChild!.getMinIntrinsicWidth(height);
 
   @override
-  bool hitTestChildren(BoxHitTestResult result, {Offset position}) =>
+  bool hitTestChildren(BoxHitTestResult result, {required Offset position}) =>
       defaultHitTestChildren(result, position: position);
 
   @override
@@ -252,13 +250,13 @@ class _ListItemRenderObject extends RenderBox
 
   @override
   void performLayout() {
-    final child = firstChild;
+    final child = firstChild!;
     final childConstraints = constraints;
     final childData = child.parentData as _ListItemData;
     child.layout(childConstraints, parentUsesSize: true);
     final childSize = child.size;
 
-    final marker = childData.nextSibling;
+    final marker = childData.nextSibling!;
     final markerConstraints = childConstraints.loosen();
     final markerData = marker.parentData as _ListItemData;
     marker.layout(markerConstraints, parentUsesSize: true);
@@ -293,7 +291,7 @@ class _ListItemRenderObject extends RenderBox
 }
 
 class _ListMarkerCircle extends _ListMarker {
-  const _ListMarkerCircle(TextStyle textStyle, {Key key})
+  const _ListMarkerCircle(TextStyle textStyle, {Key? key})
       : super(
           key: key,
           markerType: _ListMarkerType.circle,
@@ -302,7 +300,7 @@ class _ListMarkerCircle extends _ListMarker {
 }
 
 class _ListMarkerDisc extends _ListMarker {
-  const _ListMarkerDisc(TextStyle textStyle, {Key key})
+  const _ListMarkerDisc(TextStyle textStyle, {Key? key})
       : super(
           key: key,
           markerType: _ListMarkerType.disc,
@@ -311,7 +309,7 @@ class _ListMarkerDisc extends _ListMarker {
 }
 
 class _ListMarkerSquare extends _ListMarker {
-  const _ListMarkerSquare(TextStyle textStyle, {Key key})
+  const _ListMarkerSquare(TextStyle textStyle, {Key? key})
       : super(
           key: key,
           markerType: _ListMarkerType.square,
@@ -323,13 +321,13 @@ class _ListMarker extends SingleChildRenderObjectWidget {
   final _ListMarkerType markerType;
   final TextStyle textStyle;
 
-  const _ListMarker({Key key, this.markerType, this.textStyle})
+  const _ListMarker(
+      {Key? key, required this.markerType, required this.textStyle})
       : super(key: key);
 
   @override
-  RenderObject createRenderObject(BuildContext _) => _ListMarkerRenderObject()
-    ..markerType = markerType
-    ..textStyle = textStyle;
+  RenderObject createRenderObject(BuildContext _) =>
+      _ListMarkerRenderObject(markerType, textStyle);
 
   @override
   void updateRenderObject(
@@ -341,6 +339,8 @@ class _ListMarker extends SingleChildRenderObjectWidget {
 }
 
 class _ListMarkerRenderObject extends RenderBox {
+  _ListMarkerRenderObject(this._markerType, this._textStyle);
+
   _ListMarkerType _markerType;
   set markerType(_ListMarkerType v) {
     if (v == _markerType) return;
@@ -348,13 +348,12 @@ class _ListMarkerRenderObject extends RenderBox {
     markNeedsLayout();
   }
 
-  TextPainter __textPainter;
+  TextPainter? __textPainter;
   TextPainter get _textPainter {
-    __textPainter ??= TextPainter(
+    return __textPainter ??= TextPainter(
       text: TextSpan(style: _textStyle, text: '1.'),
       textDirection: TextDirection.ltr,
     )..layout();
-    return __textPainter;
   }
 
   TextStyle _textStyle;
@@ -382,14 +381,14 @@ class _ListMarkerRenderObject extends RenderBox {
     final center = offset +
         Offset(
           size.width / 2,
-          (m?.descent?.isFinite == true && m?.unscaledAscent?.isFinite == true)
+          (m?.descent.isFinite == true && m?.unscaledAscent.isFinite == true)
               ? size.height -
-                  m.descent -
+                  m!.descent -
                   m.unscaledAscent +
                   m.unscaledAscent * .7
               : size.height / 2,
         );
-    final radius = _textStyle.fontSize * .2;
+    final radius = _textStyle.fontSize! * .2;
 
     switch (_markerType) {
       case _ListMarkerType.circle:
@@ -397,7 +396,7 @@ class _ListMarkerRenderObject extends RenderBox {
           center,
           radius * .9,
           Paint()
-            ..color = _textStyle.color
+            ..color = _textStyle.color!
             ..strokeWidth = 1
             ..style = PaintingStyle.stroke,
         );
@@ -406,13 +405,13 @@ class _ListMarkerRenderObject extends RenderBox {
         canvas.drawCircle(
           center,
           radius,
-          Paint()..color = _textStyle.color,
+          Paint()..color = _textStyle.color!,
         );
         break;
       case _ListMarkerType.square:
         canvas.drawRect(
           Rect.fromCircle(center: center, radius: radius * .8),
-          Paint()..color = _textStyle.color,
+          Paint()..color = _textStyle.color!,
         );
         break;
     }
