@@ -8,14 +8,13 @@ class TagRuby {
   final BuildMetadata rubyMeta;
   final WidgetFactory wf;
 
-  BuildOp _rubyOp;
-  BuildOp _rtOp;
+  BuildOp? _rubyOp;
+  BuildOp? _rtOp;
 
   TagRuby(this.wf, this.rubyMeta);
 
   BuildOp get op {
-    _rubyOp ??= BuildOp(onChild: onChild, onTree: onTree);
-    return _rubyOp;
+    return _rubyOp ??= BuildOp(onChild: onChild, onTree: onTree);
   }
 
   void onChild(BuildMetadata childMeta) {
@@ -30,14 +29,15 @@ class TagRuby {
         _rtOp ??= BuildOp(
           onTree: (rtMeta, rtTree) {
             if (rtTree.isEmpty) return;
-            final rtBit = _RtBit(rtTree, rtTree.tsb, rtMeta, rtTree.copyWith());
+            final rtBit = _RtBit(
+                rtTree, rtTree.tsb, rtMeta, rtTree.copyWith() as BuildTree);
             rtTree.replaceWith(rtBit);
           },
         );
 
         childMeta
           ..[kCssFontSize] = '0.5em'
-          ..register(_rtOp);
+          ..register(_rtOp!);
         break;
     }
   }
@@ -45,7 +45,7 @@ class TagRuby {
   void onTree(BuildMetadata _, BuildTree tree) {
     final rubyBits = <BuildBit>[];
     for (final bit in tree.bits.toList(growable: false)) {
-      if (rubyBits.isEmpty && bit.tsb == null) {
+      if (rubyBits.isEmpty && bit is WhitespaceBit) {
         // the first bit is whitespace, just ignore it
         continue;
       }
@@ -54,7 +54,7 @@ class TagRuby {
         continue;
       }
 
-      final rtBit = bit as _RtBit;
+      final rtBit = bit;
       final rtTree = rtBit.tree;
       final rubyTree = tree.sub();
       final placeholder = WidgetPlaceholder<List<BuildTree>>([rubyTree, rtTree])
@@ -70,7 +70,7 @@ class TagRuby {
         });
 
       final anchor = rubyBits.first;
-      WidgetBit.inline(anchor.parent, placeholder).insertBefore(anchor);
+      WidgetBit.inline(anchor.parent!, placeholder).insertBefore(anchor);
 
       for (final rubyBit in rubyBits) {
         rubyTree.add(rubyBit.copyWith(parent: rubyTree));
@@ -93,6 +93,6 @@ class _RtBit extends BuildBit<Null, BuildTree> {
   BuildTree buildBit(Null _) => tree;
 
   @override
-  BuildBit copyWith({BuildTree parent, TextStyleBuilder tsb}) =>
-      _RtBit(parent ?? this.parent, tsb ?? this.tsb, meta, tree);
+  BuildBit copyWith({BuildTree? parent, TextStyleBuilder? tsb}) =>
+      _RtBit(parent ?? this.parent!, tsb ?? this.tsb, meta, tree);
 }
