@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
-import 'package:network_image_mock/network_image_mock.dart';
+import 'package:mocktail_image_network/mocktail_image_network.dart';
 
 import '_.dart';
 
@@ -347,7 +347,7 @@ void main() {
     testWidgets('renders img with sizing', (WidgetTester tester) async {
       final src = 'https://domain.com/image.jpg';
       final html = 'Foo <img src="$src" style="width: 10px; height: 10px;" />';
-      final explained = await mockNetworkImagesFor(() => explain(tester, html));
+      final explained = await mockNetworkImages(() => explain(tester, html));
       expect(
           explained,
           equals('[RichText:(:Foo '
@@ -359,7 +359,7 @@ void main() {
       final src = 'https://domain.com/image.jpg';
       final html = '<div style="text-align: center">'
           '<img src="$src" width="10" height="10" style="vertical-align: middle" /></div>';
-      final explained = await mockNetworkImagesFor(() => explain(tester, html));
+      final explained = await mockNetworkImages(() => explain(tester, html));
       expect(
           explained,
           equals('[_TextAlignBlock:child=[RichText:align=center,'
@@ -430,7 +430,7 @@ void main() {
               );
 
               await screenMatchesGolden(tester, testCase.key);
-            }, skip: null);
+            }, skip: goldenSkip != null);
           }
         }, skip: goldenSkip);
       },
@@ -459,7 +459,7 @@ void main() {
             );
 
             await screenMatchesGolden(tester, 'width');
-          }, skip: null);
+          }, skip: goldenSkip != null);
 
           testGoldens('height', (tester) async {
             await tester.pumpWidgetBuilder(
@@ -476,7 +476,7 @@ void main() {
             );
 
             await screenMatchesGolden(tester, 'height');
-          }, skip: null);
+          }, skip: goldenSkip != null);
         }, skip: goldenSkip);
       },
       config: GoldenToolkitConfiguration(
@@ -490,7 +490,7 @@ void main() {
 class _Golden extends StatelessWidget {
   final String html;
 
-  const _Golden(this.html, {Key key}) : super(key: key);
+  const _Golden(this.html, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext _) => Scaffold(
