@@ -19,7 +19,7 @@ class StyleVerticalAlign {
 
   BuildOp get buildOp => BuildOp(
         onTree: (meta, tree) {
-          if (meta.willBuildSubtree) return;
+          if (meta.willBuildSubtree == true) return;
 
           final v = meta[kCssVerticalAlign];
           if (v == null || v == kCssVerticalAlignBaseline) return;
@@ -48,7 +48,7 @@ class StyleVerticalAlign {
           tree.replaceWith(WidgetBit.inline(tree, built, alignment: alignment));
         },
         onWidgets: (meta, widgets) {
-          if (_skipBuilding[meta] == true || widgets?.isNotEmpty != true) {
+          if (_skipBuilding[meta] == true || widgets.isEmpty) {
             return widgets;
           }
 
@@ -59,8 +59,9 @@ class StyleVerticalAlign {
           return listOrNull(wf
               .buildColumnPlaceholder(meta, widgets)
               ?.wrapWith((context, child) {
-            final tsh = meta.tsb().build(context);
+            final tsh = meta.tsb.build(context);
             final alignment = _tryParseAlignmentGeometry(tsh.textDirection, v);
+            if (alignment == null) return child;
             return wf.buildAlign(meta, child, alignment);
           }));
         },
@@ -68,7 +69,7 @@ class StyleVerticalAlign {
         priority: kPriority4500,
       );
 
-  WidgetPlaceholder _buildTree(BuildMetadata meta, BuildTree tree) {
+  WidgetPlaceholder? _buildTree(BuildMetadata meta, BuildTree tree) {
     final bits = tree.bits.toList(growable: false);
     if (bits.length == 1) {
       final firstBit = bits.first;
@@ -83,23 +84,25 @@ class StyleVerticalAlign {
     return wf.buildColumnPlaceholder(meta, copied.build());
   }
 
-  Widget _buildStack(BuildContext context, BuildMetadata meta, Widget child,
+  Widget? _buildStack(BuildContext context, BuildMetadata meta, Widget child,
       EdgeInsets padding) {
-    final tsh = meta.tsb().build(context);
-    final fontSize = tsh.style.fontSize;
+    final tsh = meta.tsb.build(context);
+    final fontSize = tsh.style.fontSize!;
+    final withPadding = wf.buildPadding(
+      meta,
+      Opacity(child: child, opacity: 0),
+      EdgeInsets.only(
+        bottom: fontSize * padding.bottom,
+        top: fontSize * padding.top,
+      ),
+    );
+    if (withPadding == null) return null;
 
     return wf.buildStack(
       meta,
       tsh,
       <Widget>[
-        wf.buildPadding(
-          meta,
-          Opacity(child: child, opacity: 0),
-          EdgeInsets.only(
-            bottom: fontSize * padding.bottom,
-            top: fontSize * padding.top,
-          ),
-        ),
+        withPadding,
         Positioned(
           child: child,
           bottom: padding.top > 0 ? null : 0,
@@ -110,7 +113,7 @@ class StyleVerticalAlign {
   }
 }
 
-AlignmentGeometry _tryParseAlignmentGeometry(TextDirection dir, String value) {
+AlignmentGeometry? _tryParseAlignmentGeometry(TextDirection dir, String value) {
   final isLtr = dir != TextDirection.rtl;
   switch (value) {
     case kCssVerticalAlignTop:
@@ -126,7 +129,7 @@ AlignmentGeometry _tryParseAlignmentGeometry(TextDirection dir, String value) {
   return null;
 }
 
-PlaceholderAlignment _tryParsePlaceholderAlignment(String value) {
+PlaceholderAlignment? _tryParsePlaceholderAlignment(String value) {
   switch (value) {
     case kCssVerticalAlignTop:
     case kCssVerticalAlignSub:
