@@ -34,11 +34,11 @@ class TagImg {
         },
         onTree: (meta, tree) {
           final data = _parse(meta);
-          final built = _build(meta, data);
+          final built = wf.buildImage(meta, data);
           if (built == null) {
             final imgText = data.alt ?? data.title;
             if (imgText?.isNotEmpty == true) {
-              tree.addText(imgText);
+              tree.addText(imgText!);
             }
             return;
           }
@@ -46,33 +46,15 @@ class TagImg {
           final placeholder =
               WidgetPlaceholder<ImageMetadata>(data, child: built);
 
-          tree.replaceWith(meta.willBuildSubtree
+          tree.replaceWith(meta.willBuildSubtree!
               ? WidgetBit.block(tree, placeholder)
               : WidgetBit.inline(tree, placeholder));
         },
       );
 
-  Widget _build(BuildMetadata meta, ImageMetadata data) {
-    final source = data.sources?.first;
-    if (source == null) return null;
-
-    final provider = wf.imageProvider(source);
-    if (provider == null) return null;
-
-    var built = wf.buildImage(meta, provider, data);
-
-    if (source.height?.isNegative == false &&
-        source.width?.isNegative == false &&
-        source.height != 0) {
-      built = wf.buildAspectRatio(meta, built, source.width / source.height);
-    }
-
-    return built;
-  }
-
   ImageMetadata _parse(BuildMetadata meta) {
     final attrs = meta.element.attributes;
-    final url = wf.urlFull(attrs[kAttributeImgSrc]);
+    final url = wf.urlFull(attrs[kAttributeImgSrc] ?? '');
     return ImageMetadata(
       alt: attrs[kAttributeImgAlt],
       sources: (url != null)
@@ -83,7 +65,7 @@ class TagImg {
                 width: tryParseDoubleFromMap(attrs, kAttributeImgWidth),
               ),
             ]
-          : null,
+          : const [],
       title: attrs[kAttributeImgTitle],
     );
   }
