@@ -145,7 +145,7 @@ class _BuildOpPriority extends WidgetFactory {
   final int a;
   final int b;
 
-  _BuildOpPriority({this.a, this.b});
+  _BuildOpPriority({required this.a, required this.b});
 
   @override
   void parse(BuildMetadata meta) {
